+++ crates/lumina-video-core/src/linux_video_gst.rs
@@ -30,12 +30,12 @@ use gstreamer::prelude::*;
 use gstreamer_app as gst_app;
 use gstreamer_video as gst_video;
 
-use crate::media::{
+use crate::video::{
     CpuFrame, DecodedFrame, HwAccelType, PixelFormat, Plane, VideoDecoderBackend, VideoError,
     VideoFrame, VideoMetadata,
 };
 
-use crate::media::video::{DmaBufPlane, LinuxGpuSurface};
+use crate::video::{DmaBufPlane, LinuxGpuSurface};
 
 /// Shared audio state for GStreamer audio control.
 /// This is used to control volume/mute from the UI thread.
