+++ crates/lumina-video-core/src/macos_video.rs
@@ -80,12 +80,12 @@ use objc2_foundation::{
     NSCopying, NSMutableDictionary, NSNumber, NSObjectProtocol, NSString, NSURL,
 };
 
-use super::video::{
+use crate::video::{
     CpuFrame, DecodedFrame, HwAccelType, PixelFormat, Plane, VideoDecoderBackend, VideoError,
     VideoFrame, VideoMetadata,
 };
 
-use super::video::MacOSGpuSurface;
+use crate::video::MacOSGpuSurface;
 
 // FFI declarations for IOSurface
 extern "C" {
